user authentication and restrictions

+++ study/views.py
@@ -1,16 +1,21 @@
 from django.shortcuts import  render , redirect
 from django.db.models import Q
 from django.http import HttpResponse
-from .models import Room, Topic
-from .forms import RoomForm
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate , login , logout
+from django.contrib.auth.decorators import login_required
 from django.contrib import messages
+from django.http import HttpResponse
 
+from .models import Room, Topic
+from .forms import RoomForm
 # Create your views here.
 
 
 def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
 
     if request.method=="POST":
         username=request.POST.get('username')
@@ -57,7 +62,7 @@ def room(request,pk):
     context = {'room':room_l}
     return render(request,'study/room.html',context)
 
-
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST' :
@@ -69,10 +74,15 @@ def createRoom(request):
     context={'form': form}
     return render(request, 'study/room_form.html',context)
 
+
+@login_required(login_url='login')
 def updateRoom( request , pk):
     room = Room.objects.get(id=pk)
 
     form= RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed Here!!!")
 
     if request.method == 'POST' :
         form= RoomForm(request.POST,instance=room)
@@ -85,9 +95,12 @@ def updateRoom( request , pk):
     return render(request, 'study/room_form.html',context)
 
 
-
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse("You are not allowed Here!!!")
+
     if request.method == "POST":
         room.delete()
         return redirect('home')
